l10: restore kheap (kmalloc/kfree) in 64-bit + (uintptr_t)ptr fix in heap.c (Lecture 10)
PeachOS64BitCourse commit cf74370 ported. After this commit
kernel_main can allocate at runtime.

What changed:
  src/kernel.asm:
    PD_Table moves from 3 hand-rolled 2-MiB leaves to a NASM
    %rep 65 macro covering 130 MiB total (well past the kheap
    body at 0x01000000 + 100 MiB).

  src/kernel.c:
    + #include memory/heap/kheap.h + memory/memory.h
    + kheap_init();
    + char* data = kmalloc(50); data = ABC\0; print(data);

  src/memory/heap/heap.c (port fix):
    (unsigned int)ptr -> (uintptr_t)ptr in heap_validate_alignment.
    Caught by -Werror=pointer-to-int-cast on first build: unsigned
    int is 4 bytes on x86_64 but void* is 8, the cast was
    truncating high pointer bits. Same fix PeachOS64 makes at L10.

  src/memory/memory.c, src/memory/heap/kheap.c:
    Unchanged from 32-bit SamOs; compile clean under x86_64-elf-gcc.

  Makefile: FILES += memory.o, heap.o, kheap.o + their build rules.
  build.sh: mkdir -p build/memory build/memory/heap.

Verified by pmemsave 0xb8000 after boot:
  Hello 64-bit!
  ABC

Both strings present = kheap_init survived + kmalloc returned a
real pointer + the write/print round-tripped.

Test: tests64/L10-kheap-online.sh greps both tokens.

Doc: docs/64bit/chapters/L10-restore-kheap.md - theory on the
block-table allocator, the pointer-width port fix explained, why
PD_Table grew to 65 entries, what kheap_init does, predicted
kmalloc addresses, and a forward-looking note that paging.c has
the same truncation waiting for L11.

Next: Lecture 11 switches the paging granularity from 2 MiB to
4 KiB pages.

+++ src/memory/heap/heap.c
@@ -28,7 +28,11 @@ out:
 }
 
 static bool heap_validate_alignment(void* ptr){
-    return ((unsigned int)ptr % SAMOS_HEAP_BLOCK_SIZE) == 0;
+    // L10 port fix: `unsigned int` is 4 bytes on x86_64; casting a
+    // void* through it truncates the high 32 bits and breaks the
+    // alignment check for any pointer above 4 GiB. Use uintptr_t
+    // which is exactly the platform pointer width (8 bytes here).
+    return ((uintptr_t)ptr % SAMOS_HEAP_BLOCK_SIZE) == 0;
 }
 
 int heap_create(struct heap* heap, void* ptr, void* end, struct heap_table* table){
